# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/fund/records": {
            "get": {
                "tags": [
                    "fund_record_controller"
                ],
                "responses": {}
            },
            "post": {
                "tags": [
                    "fund_record_controller"
                ],
                "parameters": [
                    {
                        "description": "FundRecordDTO",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FundRecordDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "model.FundRecordDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "type": {
                    "type": "integer"
                }
            }
        }
    }
}